Huge performance increases with vulkan buffer uploads.

+++ Engine/Source/Runtime/Renderer/RendererUtils.h
@@ -19,7 +19,7 @@ namespace Lumina::RenderUtils
      * @param DesiredSize New desired size.
      * @return true if the buffer was resized.
      */
-    RUNTIME_API bool ResizeBufferIfNeeded(FRHIBufferRef& Buffer, uint32 DesiredSize, int GrowthFactor);
+    RUNTIME_API bool ResizeBufferIfNeeded(FRHIBufferRef& Buffer, uint32 DesiredSize, float GrowthFactor);
     
     RUNTIME_API FRHIImageRef CreateImageFromPixels(TSpan<uint8> PixelData, bool bFlipVertically = true, glm::uvec2 Size = {});
